unify all compiled C++ into 1 binary file
