+++ src/pyRF24.cpp
@@ -1,7 +1,41 @@
 #include <pybind11/pybind11.h>
 #include "pyRF24.h"
 
-PYBIND11_MODULE(rf24, m)
+void throw_ba_exception(void)
+{
+    PyErr_SetString(PyExc_TypeError, "buf parameter must be bytes or bytearray");
+    py::error_already_set();
+}
+
+char* get_bytes_or_bytearray_str(py::object buf)
+{
+    PyObject* py_ba;
+    py_ba = buf.ptr();
+    if (PyByteArray_Check(py_ba))
+        return PyByteArray_AsString(py_ba);
+    else if (PyBytes_Check(py_ba))
+        return PyBytes_AsString(py_ba);
+    else
+        throw_ba_exception();
+
+    return NULL;
+}
+
+int get_bytes_or_bytearray_ln(py::object buf)
+{
+    PyObject* py_ba;
+    py_ba = buf.ptr();
+    if (PyByteArray_Check(py_ba))
+        return PyByteArray_Size(py_ba);
+    else if (PyBytes_Check(py_ba))
+        return PyBytes_Size(py_ba);
+    else
+        throw_ba_exception();
+
+    return 0;
+}
+
+void init_rf24(py::module& m)
 {
     m.doc() = "A Python module that wraps all RF24 C++ library's API";
     m.attr("RF24_DRIVER") = RF24_DRIVER;
@@ -68,8 +102,6 @@ PYBIND11_MODULE(rf24, m)
         .export_values();
 
     // ******************** RF24 class  **************************
-    py::options options;
-    options.disable_function_signatures();
     py::class_<RF24Wrapper>(m, "RF24")
 
         // *****************************************************************************
