+++ src/pyRF24Mesh.cpp
@@ -1,11 +1,7 @@
 #include "pyRF24Mesh.h"
 
-PYBIND11_MODULE(rf24_mesh, m)
+void init_rf24mesh(py::module& m)
 {
-    m.doc() = "A Python module that wraps the RF24Mesh C++ library's API";
-    py::options options;
-    options.disable_function_signatures();
-
     m.attr("MESH_DEFAULT_ADDRESS") = MESH_DEFAULT_ADDRESS;
     m.attr("MESH_ADDR_LOOKUP") = MESH_ADDR_LOOKUP;
     m.attr("MESH_ADDR_RELEASE") = MESH_ADDR_RELEASE;
