+++ src/pyrf24/__init__.py
@@ -1,4 +1,4 @@
-from .rf24 import (
+from .pyrf24 import (  # type: ignore
     RF24,
     RF24_CRC_DISABLED,
     RF24_CRC_8,
@@ -11,8 +11,6 @@ from .rf24 import (
     RF24_PA_HIGH,
     RF24_PA_MAX,
     RF24_DRIVER,
-)
-from .rf24_network import (
     RF24Network,
     RF24NetworkHeader,
     # RF24NetworkFrame,
@@ -29,8 +27,6 @@ from .rf24_network import (
     NETWORK_REQ_ADDRESS,
     FLAG_FAST_FRAG,
     FLAG_NO_POLL,
-)
-from .rf24_mesh import (
     RF24Mesh,
     MESH_DEFAULT_ADDRESS,
     MESH_ADDR_LOOKUP,
